requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
httpx==0.26.0
python-dotenv==1.0.0
newrelic==11.0.1
